requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
SQLAlchemy>=2.0
asyncpg
alembic
passlib[bcrypt]
python-dotenv
bcrypt>=4.0.1
argon2-cffi
aiosqlite

# test/dev
pytest
httpx
pytest-asyncio
psycopg2-binary
